refactor(svelte): carry keyboard complete-area finish payload
Collapse hasBrushDraft into brushCorners as the sole draft source of
truth, move resolveFinishBrushAction to plot-brush-finish, and let the
keyboard table own normalize + select/zoom/end routing so GGPlot only
applies the finish action.
